+++ scylla-cql/src/deserialize/mod.rs
@@ -134,7 +134,7 @@ pub(crate) mod tests {
     pub(super) static CELL1: &[u8] = &[1, 2, 3];
     pub(super) static CELL2: &[u8] = &[4, 5, 6, 7];
 
-    pub(super) fn serialize_cells(
+    pub(crate) fn serialize_cells(
         cells: impl IntoIterator<Item = Option<impl AsRef<[u8]>>>,
     ) -> Bytes {
         let mut bytes = BytesMut::new();
