+++ scylla-cql/src/deserialize/row_tests.rs
@@ -1,6 +1,5 @@
 use assert_matches::assert_matches;
 use bytes::Bytes;
-use scylla_macros::DeserializeRow;
 
 use crate::deserialize::row::BuiltinDeserializationErrorKind;
 use crate::deserialize::{DeserializationError, FrameSlice, value};
@@ -79,244 +78,6 @@ fn test_deserialization_as_column_iterator() {
     assert!(col3.slice.is_none());
 
     assert!(iter.next().is_none());
-}
-
-// Do not remove. It's not used in tests but we keep it here to check that
-// we properly ignore warnings about unused variables, unnecessary `mut`s
-// etc. that usually pop up when generating code for empty structs.
-#[derive(DeserializeRow)]
-#[scylla(crate = crate)]
-struct TestUdtWithNoFieldsUnordered {}
-
-#[derive(DeserializeRow)]
-#[scylla(crate = crate, flavor = "enforce_order")]
-struct TestUdtWithNoFieldsOrdered {}
-
-// If deserialize is never called, rust warns that the struct is never constructed.
-// We don't want to `expect(dead_code)` on struct definitions, because that could silence
-// some warnings that this test is supposed to prevent.
-#[expect(unreachable_code, dead_code)]
-fn dummy_deserialize_udts() {
-    let _ = deserialize::<TestUdtWithNoFieldsUnordered>(todo!(), todo!()).unwrap();
-    let _ = deserialize::<TestUdtWithNoFieldsOrdered>(todo!(), todo!()).unwrap();
-}
-
-#[test]
-fn test_struct_deserialization_loose_ordering() {
-    #[derive(DeserializeRow, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate")]
-    struct MyRow<'a> {
-        a: &'a str,
-        b: Option<i32>,
-        #[scylla(skip)]
-        c: String,
-        #[scylla(default_when_null)]
-        d: i32,
-        #[scylla(default_when_null)]
-        e: &'a str,
-    }
-
-    // Original order of columns
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-        spec("d", ColumnType::Native(NativeType::Int)),
-        spec("e", ColumnType::Native(NativeType::Text)),
-    ];
-    let byts = serialize_cells([val_str("abc"), val_int(123), None, val_str("def")]);
-    let row = deserialize::<MyRow<'_>>(specs, &byts).unwrap();
-    assert_eq!(
-        row,
-        MyRow {
-            a: "abc",
-            b: Some(123),
-            c: String::new(),
-            d: 0,
-            e: "def",
-        }
-    );
-
-    // Different order of columns - should still work
-    let specs = &[
-        spec("e", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-        spec("d", ColumnType::Native(NativeType::Int)),
-        spec("a", ColumnType::Native(NativeType::Text)),
-    ];
-    let byts = serialize_cells([None, val_int(123), None, val_str("abc")]);
-    let row = deserialize::<MyRow<'_>>(specs, &byts).unwrap();
-    assert_eq!(
-        row,
-        MyRow {
-            a: "abc",
-            b: Some(123),
-            c: String::new(),
-            d: 0,
-            e: "",
-        }
-    );
-
-    // Missing column
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("e", ColumnType::Native(NativeType::Text)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-
-    // Missing both default_when_null column
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-
-    // Wrong column type
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Int)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-}
-
-#[test]
-fn test_struct_deserialization_strict_ordering() {
-    #[derive(DeserializeRow, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate", flavor = "enforce_order")]
-    struct MyRow<'a> {
-        a: &'a str,
-        b: Option<i32>,
-        #[scylla(skip)]
-        c: String,
-        #[scylla(default_when_null)]
-        d: i32,
-        #[scylla(default_when_null)]
-        e: &'a str,
-    }
-
-    // Correct order of columns
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-        spec("d", ColumnType::Native(NativeType::Int)),
-        spec("e", ColumnType::Native(NativeType::Text)),
-    ];
-    let byts = serialize_cells([val_str("abc"), val_int(123), None, val_str("def")]);
-    let row = deserialize::<MyRow<'_>>(specs, &byts).unwrap();
-    assert_eq!(
-        row,
-        MyRow {
-            a: "abc",
-            b: Some(123),
-            c: String::new(),
-            d: 0,
-            e: "def",
-        }
-    );
-
-    // Wrong order of columns
-    let specs = &[
-        spec("b", ColumnType::Native(NativeType::Int)),
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("d", ColumnType::Native(NativeType::Int)),
-        spec("e", ColumnType::Native(NativeType::Text)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-
-    // Missing column
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("e", ColumnType::Native(NativeType::Text)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-
-    // Missing both default_when_null column
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-
-    // Wrong column type
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Int)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-    ];
-    MyRow::type_check(specs).unwrap_err();
-}
-
-#[test]
-fn test_struct_deserialization_no_name_check() {
-    #[derive(DeserializeRow, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate", flavor = "enforce_order", skip_name_checks)]
-    struct MyRow<'a> {
-        a: &'a str,
-        b: Option<i32>,
-        #[scylla(skip)]
-        c: String,
-    }
-
-    // Correct order of columns
-    let specs = &[
-        spec("a", ColumnType::Native(NativeType::Text)),
-        spec("b", ColumnType::Native(NativeType::Int)),
-    ];
-    let byts = serialize_cells([val_str("abc"), val_int(123)]);
-    let row = deserialize::<MyRow<'_>>(specs, &byts).unwrap();
-    assert_eq!(
-        row,
-        MyRow {
-            a: "abc",
-            b: Some(123),
-            c: String::new(),
-        }
-    );
-
-    // Correct order of columns, but different names - should still succeed
-    let specs = &[
-        spec("z", ColumnType::Native(NativeType::Text)),
-        spec("x", ColumnType::Native(NativeType::Int)),
-    ];
-    let byts = serialize_cells([val_str("abc"), val_int(123)]);
-    let row = deserialize::<MyRow<'_>>(specs, &byts).unwrap();
-    assert_eq!(
-        row,
-        MyRow {
-            a: "abc",
-            b: Some(123),
-            c: String::new(),
-        }
-    );
-}
-
-#[test]
-fn test_struct_deserialization_cross_rename_fields() {
-    #[derive(scylla_macros::DeserializeRow, PartialEq, Eq, Debug)]
-    #[scylla(crate = crate)]
-    struct TestRow {
-        #[scylla(rename = "b")]
-        a: i32,
-        #[scylla(rename = "a")]
-        b: String,
-    }
-
-    // Columns switched wrt fields - should still work.
-    {
-        let row_bytes =
-            serialize_cells(["The quick brown fox".as_bytes(), &42_i32.to_be_bytes()].map(Some));
-        let specs = [
-            spec("a", ColumnType::Native(NativeType::Text)),
-            spec("b", ColumnType::Native(NativeType::Int)),
-        ];
-
-        let row = deserialize::<TestRow>(&specs, &row_bytes).unwrap();
-        assert_eq!(
-            row,
-            TestRow {
-                a: 42,
-                b: "The quick brown fox".to_owned(),
-            }
-        );
-    }
 }
 
 fn val_int(i: i32) -> Option<Vec<u8>> {
@@ -1002,20 +763,6 @@ fn metadata_does_not_bound_deserialized_rows() {
             col_spec("text", ColumnType::Native(NativeType::Text)),
         ];
 
-        // Tuple
-        let decoded_tuple_res = deserialize::<(&[u8], &str)>(row_typ, &bytes);
-
-        // Custom struct
-        #[derive(DeserializeRow)]
-        #[scylla(crate=crate)]
-        struct MyRow<'frame> {
-            #[expect(dead_code)]
-            bytes: &'frame [u8],
-            #[expect(dead_code)]
-            text: &'frame str,
-        }
-        let decoded_custom_struct_res = deserialize::<MyRow>(row_typ, &bytes);
-
-        (decoded_tuple_res, decoded_custom_struct_res)
+        deserialize::<(&[u8], &str)>(row_typ, &bytes)
     };
 }
